refactor(hero): convert to RSC with client islands for animations
Restructures the 444-line all-'use client' Hero into:

- Hero.tsx (RSC): renders BOTH .hero--desktop and .hero--mobile
  variants with all static content. LCP element (p.hero__tagline) is
  now static SSR markup; no hydration on the critical path.
- HeroBootAnimation.tsx (new): wraps the bootRef + runBoot() call.
  Each variant mounts its own instance with a variant prop
  ('desktop' | 'mobile'); useEffect gates on
  window.matchMedia('(max-width: 768px)').matches and the hidden
  variant's instance no-ops. matchMedia is deterministic regardless
  of stylesheet load order -- chosen over getComputedStyle to avoid
  the hydration race documented in spec section 6.
- HeroSystemFailure.tsx (new): wraps the sysfail overlay + event
  listeners. Listens for 'hero:sysfail:show' / 'hero:sysfail:hide'
  window events dispatched by HeroBootAnimation; decoupled from boot
  animation via custom events rather than shared refs.
- _layout.css: new .hero--desktop / .hero--mobile media-query toggle
  hides the non-matching variant at max-width: 768px.

useBreakpoint is no longer imported by Hero (still used in AppShell
for chrome decisions). runBoot moved to HeroBootAnimation and exported
for the existing sysfail-loop unit test.

Trade-offs: both variants in DOM doubles Hero markup bytes (~5KB to
~10KB HTML); acceptable since both are pure RSC (no hydration cost
per hidden variant). Sysfail now communicates via window events
instead of direct ref manipulation -- same observable behavior, one
extra event dispatch per trigger.

Updates hero-heading.test.ts and sysfail-loop.test.ts for the
restructured component and moved runBoot export.

Implements Fix C of spec docs/superpowers/specs/2026-05-18-mobile-
lcp-perf-fix-design.md.

Reversal: revert Hero.tsx; delete the two island files; revert the
_layout.css toggle; revert the two updated tests.
